animal main image & add images
- define 'main' image for an animal @ EditAnimalForm
- add images / additional images on EditAnimalForm

+++ src/components/Animal/AnimalList.jsx
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
 import axios from 'axios';
+import placeholder from '../../assets/placeholder.png'
 
 function AnimalList() {
 
@@ -12,7 +13,7 @@ function AnimalList() {
     // get data of all animals in DB
     axios.get('http://localhost:5001/allAnimals')
       .then((res) => {
-setState('success');
+        setState('success');
         console.log(res.data)
         setAnimals(res.data);
               })
@@ -24,10 +25,17 @@ setState('success');
 
   // use id passed in onclick function to delete the animal w/that id from the DB
   function deleteAnimal(id) {
-    axios.delete(`http://localhost:5001/delete-animal/${id}`)
-    alert('animal supprimé')
-    // reload page to update list
-    window.location.reload()
+    // confirm action & warn user that this action cannot be undone
+    const confirmed = window.confirm(`Êtes-vous sûr de vouloir supprimer cet animal ?\nCette action ne peut pas être inversé...`)
+
+    // if user confirms deletion, remove user entry from DB
+    if (confirmed) {
+      axios.delete(`http://localhost:5001/delete-animal/${id}`)
+      alert('animal supprimé')
+      // reload page to update list
+      setState('loading');
+      window.location.reload()
+    }
   }
 
   while (state === 'loading') {
@@ -53,8 +61,9 @@ setState('success');
       <table>
         <thead>
         <tr>
-          <th>ICAD</th>
-          <th>Nom</th>
+            <th></th>
+            <th>Nom</th>
+            <th>ICAD</th>
           <th>Sexe</th>
           <th>Race</th>
           <th>Actions</th>
@@ -63,8 +72,15 @@ setState('success');
         <tbody>
         {animals.map((animal) => (
           <tr key={animal._id}>
-            <td>{animal.numICAD}</td>
+            <td>
+              <div className="img-list-icon">{
+                animal.images.length > 0 ?
+                  <img src={`http://localhost:5001/${animal.images[0]}`} alt={animal.name} />
+                  : <img src={placeholder} alt='none' />
+              }</div>
+            </td>
             <td>{animal.name}</td>
+            <td>{animal.numICAD}</td>
             <td>{animal.sex}</td>
             <td>{animal.race}</td>
             <td>
